+++ src/config/i18n.js
@@ -3,21 +3,13 @@ module.exports = {
   langs:{
     en: {
       path: `en`,
-      locale: `en-US`,
-      dateFormat: `DD/MM/YYYY`,
-      siteLanguage: `en`,
-      ogLanguage: `en_US`,
-      defaultTitle: `Using i18n with Gatsby`,
-      defaultDescription: `Gatsby example site using MDX and dependency-free i18n`,
+      name: 'English',
+      countryCode: 'GB',
     },
     ja: {
       path: `ja`,
-      locale: `ja-JP`,
-      dateFormat: `YYYY年MM月DD日`,
-      siteLanguage: `ja`,
-      ogLanguage: `ja_JP`,
-      defaultTitle: `Using i18n with Gatsby`,
-      defaultDescription: `Gatsby example site using MDX and dependency-free i18n`,
+      name: '日本語',
+      countryCode: 'JP',
     },
   }
 }
